Make libc-trusty ASLR-ready
* Use rcrt instead of crt for static PIEs
* Use new brk() interface for high addresses

Bug: 80146040

+++ lib/libc-trusty/malloc.c
@@ -55,10 +55,10 @@ void* sbrk(ptrdiff_t increment) {
     if (!__libc_brk)
         __libc_brk = (char*)_trusty_brk(0);
 
-    start = (char*)round_up((long)__libc_brk, SBRK_ALIGN);
-    end = start + round_up((long)increment, SBRK_ALIGN);
+    start = (char*)round_up((unsigned long)__libc_brk, SBRK_ALIGN);
+    end = start + round_up((unsigned long)increment, SBRK_ALIGN);
 
-    new_brk = (char*)_trusty_brk((uint32_t)(uintptr_t)end);
+    new_brk = (char*)_trusty_brk(end);
     if (new_brk < end)
         return (void*)-1;
